refactor(bigtable): make bigtable_mocks header only
